Refactor class name of QueryProcessor to MySQLProcessor
