+++ native/src/main/java/io/ballerina/stdlib/persist/Utils.java
@@ -55,17 +55,17 @@ public class Utils {
     private Utils() {
     }
 
-    static BString getEntity(Environment env) {
+    public static BString getEntity(Environment env) {
         String entity = env.getFunctionName().split("\\$")[2];
         return fromString(entity);
     }
 
-    static BObject getPersistClient(BObject client, BString entity) {
+    public static BObject getPersistClient(BObject client, BString entity) {
         BMap<?, ?> persistClients = (BMap<?, ?>) client.get(Constants.PERSIST_CLIENTS);
         return (BObject) persistClients.get(entity);
     }
 
-    static BArray[] getMetadata(RecordType recordType) {
+    public static BArray[] getMetadata(RecordType recordType) {
         ArrayType stringArrayType = TypeCreator.createArrayType(PredefinedTypes.TYPE_STRING);
 
         //TODO: use PREDEFINED TYPE.TYPE_TYPEDESC once NPE issue is resolved
@@ -125,7 +125,7 @@ public class Utils {
         return fieldsArray;
     }
 
-    static Object getFutureResult(BFuture future) {
+    public static Object getFutureResult(BFuture future) {
         while (!future.isDone()) {
             try {
                 Thread.sleep(100);
@@ -159,7 +159,7 @@ public class Utils {
         return returnArray;
     }
 
-    static Object getKey(Environment env, BArray path) {
+    public static Object getKey(Environment env, BArray path) {
         Parameter[] pathParams = env.getFunctionPathParameters();
         if (pathParams.length == 1) {
             return path.get(0);
