+++ native/src/main/java/io/ballerina/stdlib/persist/datastore/MySQLProcessor.java
@@ -16,7 +16,7 @@
  *  under the License.
  */
 
-package io.ballerina.stdlib.persist;
+package io.ballerina.stdlib.persist.datastore;
 
 import io.ballerina.runtime.api.Environment;
 import io.ballerina.runtime.api.PredefinedTypes;
@@ -31,6 +31,8 @@ import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BStream;
 import io.ballerina.runtime.api.values.BString;
 import io.ballerina.runtime.api.values.BTypedesc;
+import io.ballerina.stdlib.persist.Constants;
+import io.ballerina.stdlib.persist.ModuleUtils;
 
 import static io.ballerina.stdlib.persist.Utils.getEntity;
 import static io.ballerina.stdlib.persist.Utils.getFutureResult;
@@ -43,9 +45,9 @@ import static io.ballerina.stdlib.persist.Utils.getPersistClient;
  *
  * @since 0.5.6
  */
-public class QueryProcessor {
+public class MySQLProcessor {
 
-    private QueryProcessor() {
+    private MySQLProcessor() {
     }
 
     public static BStream query(Environment env, BObject client, BTypedesc recordType) {
